requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.116.0
frozenlist==1.7.0
gevent==25.5.1
geventhttpclient==2.3.4
greenlet==3.2.3
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.2.6
opencv-python==4.12.0.88
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-rapidjson==1.20
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tritonclient==2.59.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yarl==1.20.1
zope.event==5.1
zope.interface==7.2
